requirements.txt
aio-pika==9.5.4
python-dotenv==1.0.1
redis==5.2.1
fastapi==0.115.7
uvicorn==0.34.0